requirements.txt
alabaster==1.0.0
amqp==5.3.1
asgiref==3.8.1
attrs==24.3.0
Beaker==1.13.0
billiard==4.2.1
blinker==1.9.0
cattrs==24.1.2
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
class-registry==2.1.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cligj==0.7.2
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.5
cryptography==44.0.0
defusedxml==0.7.1
Django==5.1.5
django-allauth==65.3.1
django-celery-beat==2.7.0
django-cors-headers==4.6.0
django-debug-toolbar==5.0.1
django-extensions==3.2.3
django-filter==24.3
django-model-utils==5.0.0
django-oidc==0.1.4
django-redis==5.4.0
django-restframework==0.0.1
django-storages==1.14.4
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework-csv==3.0.2
djangorestframework-gis==1.1
et_xmlfile==2.0.0
filelock==3.16.1
filters==1.3.2
fiona==1.10.1
Flask==3.1.0
fsspec==2024.12.0
future==1.0.0
geojson==3.2.0
geos==0.2.3
gevent==24.11.1
greenlet==3.1.1
gunicorn==23.0.0
httplib2==0.22.0
huggingface-hub==0.27.1
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
itypes==1.2.0
jdcal==1.4.1
Jinja2==3.1.5
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jwt==1.3.1
kombu==5.4.2
lxml==5.3.0
Mako==1.3.8
MarkupSafe==3.0.2
msgpack-python==0.5.6
numpy==2.2.2
oauth2client==4.1.3
oauthlib==3.2.2
oic==0.10.0.0
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
postgis==1.0.4
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
py3rijndael==0.3.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodomex==3.21.0
pyjwkest==1.4.2
pyOpenSSL==25.0.0
pyparsing==3.2.1
pyproj==3.7.0
pyrsistent==0.20.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
requests-cache==1.2.1
requests-oauthlib==2.0.0
responses==0.25.6
rpds-py==0.22.3
rsa==4.9
Rtree==1.3.0
sacremoses==0.1.1
safetensors==0.5.2
sendgrid==3.6.5
sendgrid-django==4.2.0
setuptools==75.8.0
shapely==2.0.7
six==1.17.0
sqlparse==0.5.3
testfixtures==8.3.0
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.48.1
typing_extensions==4.12.2
tzdata==2024.2
unicodecsv==0.14.1
uritemplate==4.1.1
url-normalize==1.4.3
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
zipp==3.21.0
zope.event==5.0
zope.interface==7.2